Add support for logging impersonation sessions (issue #7)
This is first draft at support for storing impersonation sessions -
who was impersonating who, for how long, and when. It hooks into the
session_begin and session_end signals - creating a new ImpersonationLog
object on the former, and then closing it out (by adding the end
timestamp and session duration) on the latter.

Also included an admin list page.

--HG--
branch : impersonation-log