Downgrade signals logging from warning to debug
In production we are seeing the warnings raised by the
signals appearing frequently. I believe this is caused by
people having multiple tabs open, stopping impersonating
in one, and then some time later clicking the stop impersonate
link in another. This is a completely valid use case, and
shouldn't raise a warning.